- Metrics should now compile with GCC also.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@695 494c09c7-a67d-49d4-9938-151fb1965be8

+++ test/collision/CollisionTestSuite.cpp
@@ -103,12 +103,12 @@ void testCollisionStrategies()
             ProximityStrategyPQP::make(), workcell, device, state)
     };
 
-    for (int i = 0; i < sizeof(qs) / sizeof(*qs); i++) {
+    for (size_t i = 0; i < sizeof(qs) / sizeof(*qs); i++) {
         Q q(7);
         for (int j = 0; j < 7; j++) q[j] = qs[i][j];
 
         const bool b0 = constraints[0].getQConstraint().inCollision(q);
-        for (int j = 1; j < sizeof(constraints) / sizeof(*constraints); j++) {
+        for (size_t j = 1; j < sizeof(constraints) / sizeof(*constraints); j++) {
             const bool bJ = constraints[j].getQConstraint().inCollision(q);
 
             // This is what the test looks like:
